requirements.txt
Django==5.2.9
gunicorn==23.0.0
pillow==11.1.0
pillow_heif==0.21.0
mysqlclient==2.2.7
django-debug-toolbar==5.2.0
django-allauth[socialaccount]==0.63.3
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-recaptcha==4.0.0
qrcode==8.2
reportlab==4.4.0
bleach==6.2.0
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3
djangorestframework==3.14.0
django-cors-headers==4.3.1